moving the delete confirmation to funding file
